PS-7477: Fixing local variable binding in stored procedures
Issue: WL#9384 in 67c3c70e4 refactored how expressions are constructed,
but made the incorrect assumption that once an Item_field is created, it
will never be reassigned to another thread.
Unfortunately this isn't true for stored triggers, where Items are
reused between executions, and this change completely broke local
variables there.

Fix: as a workaround this diff adds a new method to the Item interface,
"change_thd". This is only called when a stored procedure evaluates an
expression, and reassigns the expression tree to the new Thread.